TEST: move T_Callbacks into own test suite
